set base in module
