+++ apps/api/src/gateway/gateway.service.ts
@@ -9,39 +9,23 @@ import type {
   RemoteAssignment
 } from '@open-data-capture/common/assignment';
 
-import { ConfigurationService } from '@/configuration/configuration.service';
 import { InstrumentsService } from '@/instruments/instruments.service';
 
 @Injectable()
 export class GatewayService {
-  private readonly gatewayBaseUrl: string;
   private readonly logger = new Logger(GatewayService.name);
 
   constructor(
-    configurationService: ConfigurationService,
     private readonly httpService: HttpService,
     private readonly instrumentsService: InstrumentsService
-  ) {
-    if (configurationService.get('NODE_ENV') === 'production') {
-      const internalNetworkUrl = configurationService.get('GATEWAY_INTERNAL_NETWORK_URL');
-      const siteAddress = configurationService.get('GATEWAY_SITE_ADDRESS')!;
-      if (siteAddress.hostname === 'localhost' && internalNetworkUrl) {
-        this.gatewayBaseUrl = internalNetworkUrl.href;
-      } else {
-        this.gatewayBaseUrl = siteAddress.href;
-      }
-    } else {
-      const gatewayPort = configurationService.get('GATEWAY_DEV_SERVER_PORT')!;
-      this.gatewayBaseUrl = `http://localhost:${gatewayPort}`;
-    }
-  }
+  ) {}
 
   async createRemoteAssignment(
     assignment: Assignment,
     publicKey: PublicEncryptionKey
   ): Promise<MutateAssignmentResponseBody> {
     const instrument = await this.instrumentsService.findById(assignment.instrumentId);
-    const response = await this.httpService.axiosRef.post(`${this.gatewayBaseUrl}/api/assignments`, {
+    const response = await this.httpService.axiosRef.post(`/api/assignments`, {
       ...assignment,
       instrumentBundle: instrument.bundle,
       publicKey: await publicKey.toJSON()
@@ -55,7 +39,7 @@ export class GatewayService {
   }
 
   async deleteRemoteAssignment(id: string): Promise<MutateAssignmentResponseBody> {
-    const response = await this.httpService.axiosRef.delete(`${this.gatewayBaseUrl}/api/assignments/${id}`);
+    const response = await this.httpService.axiosRef.delete(`/api/assignments/${id}`);
     if (response.status !== HttpStatus.OK) {
       throw new BadGatewayException(`Unexpected Status Code From Gateway: ${response.status}`, {
         cause: response.statusText
@@ -65,7 +49,7 @@ export class GatewayService {
   }
 
   async fetchRemoteAssignments({ subjectId }: { subjectId?: string } = {}): Promise<RemoteAssignment[]> {
-    const response = await this.httpService.axiosRef.get(`${this.gatewayBaseUrl}/api/assignments`, {
+    const response = await this.httpService.axiosRef.get(`/api/assignments`, {
       params: {
         subjectId
       }
